chore(sc2bot): more ruff micro-cleanups (round 15)
Auto-fixed 11 sites:

- RUF046 (7): drop unnecessary int(int_value) / int(round(x)) double-casts.
- RUF015 (1): list(iter)[0] → next(iter(iter)) — avoids materializing
  the whole list when you only want the first element.
- PIE810 (3): startswith()/endswith() chains over the same string can
  use a tuple argument instead — e.g. s.startswith("a") or
  s.startswith("b") → s.startswith(("a", "b")).

Verified: pytest 372 passed / 35 skipped / 0 failed.

+++ wicked_zerg_challenger/utils/spatial_partition.py
@@ -39,8 +39,8 @@ class SpatialGrid:
         self.map_height = map_size[1]
 
         # Calculate grid dimensions
-        self.grid_width = int(math.ceil(self.map_width / self.cell_size))
-        self.grid_height = int(math.ceil(self.map_height / self.cell_size))
+        self.grid_width = math.ceil(self.map_width / self.cell_size)
+        self.grid_height = math.ceil(self.map_height / self.cell_size)
 
         # Grid storage: cell_key -> list of (position, data)
         self.grid: dict[tuple[int, int], list[tuple[tuple[float, float], Any]]] = {}
@@ -136,7 +136,7 @@ class SpatialGrid:
 
         # Calculate cells to check
         cells_to_check = (
-            int(math.ceil(radius / self.cell_size)) + 1
+            math.ceil(radius / self.cell_size) + 1
             if self.cell_size
             else int(radius) + 1
         )
